Use correct namespace for derived SpatialQueriesFixtures
