+++ Tests.NHibernate.Spatial.PostGis/PostGisSpatialQueriesFixture.cs
@@ -1,8 +1,9 @@
 using NHibernate;
 using NHibernate.Cfg;
 using NUnit.Framework;
+using Tests.NHibernate.Spatial.RandomGeometries;
 
-namespace Tests.NHibernate.Spatial.RandomGeometries
+namespace Tests.NHibernate.Spatial
 {
     [TestFixture]
     public class PostGisSpatialQueriesFixture : SpatialQueriesFixture
